voice_agent/requirements.txt
livekit==0.17.0
livekit-agents==0.11.0
livekit-plugins-google==0.9.0
httpx==0.27.0
python-dotenv==1.0.0
